Handle availability of fhlinkat() in FreeBSD
A function called fhlink() is defined.  It looks to be an API into
either a FreeBSD kernel module for NFS Ganesha and/or a patched
version of FreeBSD.  The API is really that of fhlinkat(), which is
available in recent FreeBSD versions.  So, detect fhlinkat() and use
it when available.

When fhlinkat() is defined then assume getfhat() and fhreadlink() are
also defined.

The approach taken here could probably be simplified by adding an
upper limit for __FreeBSD_cc_version.  However, documentation for that
variable does not seem to exist.

Addition of the relevant FreeBSD system calls is discussed at:

  https://reviews.freebsd.org/D18359

and merged at:

  https://reviews.freebsd.org/rS341689

However, there are no clues indicating what FreeBSD version they were
merged for.

This could possibly be resolved by cloning the FreeBSD repository and
doing more research there.  Sadly, this sets the bar too high for
casual user-space contributions.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: I700a8e11f0c2c2f9efcf63e0a8256aa7a0407c29