scripts: Add -brief and -paths options to manager_mgr show exports
Given that paths can contain spaces and commas, the regular output
format of "ganesha_mgr show exports" isn't terribly useful for parsing
by something else.

CTDB wants to monitor exported paths.  Parsing ganesha.conf in a shell
script is possible but can't be done in a readable, maintainable way.

So, add a couple of options that make it easier to parse out the
paths.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: Ibdc8c66495a16f6c5526ec4b752a4e1add513724