D-Bus: Drop default recovery/grace event, improve argument checking
If just an IP address is supplied then the default event is TAKE_IP.
However, this causes ambiguity when used with IPv6 due to the use of
colon (':') as a separator.

For example, 2:ab:cd::42 can logically mean 2 completely different
things:

* Event 2 (RELEASE_IP) with IP address ab:cd::42

* Default event 5 (TAKE_IP) with IP address 2:ab:cd::42

Another insane case is that specifying 0 (without a trailing colon,
attempting to mean JUST_GRACE) currently causes a TAKE_IP for IP
address 0.0.0.0.  Nobody wants this.

The behaviour is changed to make the event mandatory.  This way, the
colon (':') separator should indicate the presence of a mandatory
argument for the event.  However, this would break backward
compatibility for many users, so an empty argument (e.g. "0:") is
treated the same as a missing argument (i.e. no colon-separator).

Additionally, the use of ip and gsp.ipaddr to represent an argument
that might end up being a node ID has always hurt my brain when
reading this function.

In summary, do full event ID and associated argument checking with
less confusing variable names.

The result of the IP address parse is currently unused.  Using it
would over-complicate this commit and it can be done in an additional
commit with minimal churn.  However, the argument checking is now done
at this high level instead of dealing with it down in the depths,
where some errors are currently ignored.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: Ia03f9217385a45cb65121d966abb6e6e757abcd3